Split semaphore set operations into separate files
